feat: Adds resource detection processor (#17)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Short description of the changes

The resource detection processor is pretty popular. This adds it to the
manifest for the honeycomb collector build.

## How to verify that this has the expected result

Build it with this configuration and use the following config snippet
for processors:

```yaml
processors:
  batch:
  memory_limiter:
    check_interval: 1s
    limit_mib: 20000
    spike_limit_mib: 2000
  resourcedetection/heroku:
    detectors: [env, heroku]
    timeout: 5s
    override: false
```