chore: add a tmp volume to the dockerfile for the proguard processor (#49)
The library we are using for the proguard mapping depends on [receiving
a mapping
file](https://github.com/honeycombio/symbolic-go/blob/3589735d2fa511e5c868df821e55684c715032df/proguard.go#L19).
So we cache the proguardmapper.

Basing this change on this
[article](https://asciijungle.com/posts/2025-05-06-debugging-file-permissions-in-aws-ecs.html#fn1)

We are currently running into file permissions errors with ECS because
the proguard processor is trying to write temp files. Adding the
recommended fix.